Only delete vars.yml when it exists

+++ tasks/bosh-delete-deployment.sh
@@ -6,11 +6,13 @@ bosh2 -n -d $DEPLOYMENT_NAME delete-deployment
 
 commit_if_changed=$(readlink -f 1-click/tasks/commit-if-changed.sh)
 
-pushd state/environments/softlayer/director/$BOSH_LITE_NAME/cf-deployment
-    git rm -f vars.yml
-popd
-pushd state
-    $commit_if_changed "Update state for environments/softlayer/director/$BOSH_LITE_NAME/cf-deployment"
-popd
+if [ -e state/environments/softlayer/director/$BOSH_LITE_NAME/cf-deployment/vars.yml ]; then
+    pushd state/environments/softlayer/director/$BOSH_LITE_NAME/cf-deployment
+        git rm -f vars.yml
+    popd
+    pushd state
+        $commit_if_changed "Update state for environments/softlayer/director/$BOSH_LITE_NAME/cf-deployment"
+    popd
+fi
 
 cp -a state/. out-state
